REFACTOR Updated references to go mod github.com/haproxytech/haproxy-consul-connect
