+++ consul/logger.go
