+++ haproxy/haproxy.go
@@ -7,12 +7,12 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/criteo/haproxy-consul-connect/consul"
-	"github.com/criteo/haproxy-consul-connect/haproxy/dataplane"
-	"github.com/criteo/haproxy-consul-connect/haproxy/haproxy_cmd"
-	"github.com/criteo/haproxy-consul-connect/haproxy/state"
-	"github.com/criteo/haproxy-consul-connect/lib"
 	spoe "github.com/criteo/haproxy-spoe-go"
+	"github.com/haproxytech/haproxy-consul-connect/consul"
+	"github.com/haproxytech/haproxy-consul-connect/haproxy/dataplane"
+	"github.com/haproxytech/haproxy-consul-connect/haproxy/haproxy_cmd"
+	"github.com/haproxytech/haproxy-consul-connect/haproxy/state"
+	"github.com/haproxytech/haproxy-consul-connect/lib"
 	"github.com/hashicorp/consul/api"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	log "github.com/sirupsen/logrus"
