+++ haproxy/state/downstream.go
@@ -3,7 +3,7 @@ package state
 import (
 	"fmt"
 
-	"github.com/criteo/haproxy-consul-connect/consul"
+	"github.com/haproxytech/haproxy-consul-connect/consul"
 	"github.com/haproxytech/models"
 )
